refactor: 서비로 위치 변경
